requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
python-dotenv
phonenumbers
curl_cffi
jinja2
weasyprint
pymupdf
